To get the title to the main page from the business page is a long process.
Business content is displayed in admin or any other controller. The business content can only inform what the title of the page should be
So an event was added to the admin controller which is called before any rendering was done.

Here header title is set ( for future more things can be added) in session.

A Panel Head Controller was created to transfer the session variable to panel main twig. But the problem was that it was a string and not a template inheritence.
Thus base template was modified to include header block which was overwritten in PanelHeadController.php twig.

I can't say for now what scenarios can emerge later but for now this is a solution.