Problem with autocomplete of symfony. There is not much documentation about how to test a symfony ux autocomplete. Thus the tests had to be changed to use the URL with post and body of post as if it was a API request. Not ideal but usable.
The autocomplete also causes another problem with DTO methods that it is based on entity types and thus is validated with entity corrupting the architecture.
So we mark them unmapped and use another field to keep DTO separate from entity layers. The incoming values are mapped into the new field in events for validation

And after set data , parentId is again made into an entity for display in templates

CSRF protection is now disabled for tests in framework.yaml as using post api will cause crsf errors in form

Find function of doctrine does not like a null value to find