Search based on filter controller is providing right data. In next step it has to be processed
A DTO is created with key and value. An array is created out it and selection flag.
An array of such arrays is created.

Corresponding form is created. The filter data is fetched from the database . The filter data is formatted to be displayed in the form.
The form contains a field of collection type form elements. The output is set of values under same key.