Search is based on the filter from apply.
The flow is as follows:
The user selects a category and is presented with a filter screen ( from the home SideBarController.php)

The user selected desired filters and then presses apply.
The Url is updated with keys and category name ( which should be unique) and sent to home url with filters in the keys. The ContentController.php then sends it to ProductController.php with right parameters and search is performed and product is listed.

The search function remain in ProductController.php.

Todo: find children of hierarchy fo categories after the function has been developed